+++ tests/JoinGameForm.test.ts
@@ -3,7 +3,7 @@ import 'ts-jest';
 import { render, fireEvent } from '@testing-library/svelte';
 import { resetTestState } from './test-helper';
 import AppFixture from './AppFixture.svelte';
-import PlayerForm from '../components/PlayerForm.svelte';
+import JoinGameForm from '../components/JoinGameForm.svelte';
 import { currentPlayerId, players } from '../stores/player';
 const socket = require('socket.io-client')('test');
 
@@ -12,7 +12,7 @@ afterEach(() => {
 });
 
 test('should enable the form by default', () => {
-  const { getByText, getByLabelText } = render(AppFixture, { socket, component: PlayerForm });
+  const { getByText, getByLabelText } = render(AppFixture, { socket, component: JoinGameForm });
 
   const nameField = getByLabelText('Name') as HTMLInputElement;
   const button = getByText('Join game') as HTMLButtonElement;
@@ -22,7 +22,7 @@ test('should enable the form by default', () => {
 });
 
 test('should disable the button on submit', async () => {
-  const { getByText } = render(AppFixture, { socket, component: PlayerForm });
+  const { getByText } = render(AppFixture, { socket, component: JoinGameForm });
 
   const button = getByText('Join game') as HTMLButtonElement;
   await fireEvent.submit(button.form);
@@ -32,7 +32,7 @@ test('should disable the button on submit', async () => {
 
 test('should set the current player', async () => {
   jest.spyOn(currentPlayerId, 'set');
-  const { getByText, getByLabelText } = render(AppFixture, { socket, component: PlayerForm });
+  const { getByText, getByLabelText } = render(AppFixture, { socket, component: JoinGameForm });
 
   const nameField = getByLabelText('Name') as HTMLInputElement;
   const button = getByText('Join game') as HTMLButtonElement;
@@ -44,7 +44,7 @@ test('should set the current player', async () => {
 
 test('should add the player to the room', async () => {
   jest.spyOn(socket, 'emit');
-  const { getByText, getByLabelText } = render(AppFixture, { socket, component: PlayerForm });
+  const { getByText, getByLabelText } = render(AppFixture, { socket, component: JoinGameForm });
 
   const nameField = getByLabelText('Name') as HTMLInputElement;
   nameField.value = 'Binks';
@@ -60,7 +60,7 @@ test('should add the player to the room', async () => {
 
 test('should disable avatars that players have selected', () => {
   players['player::add']({ id: '1234', avatar: '🐶', name: 'Binks' });
-  const { getByText, getByRole } = render(AppFixture, { socket, component: PlayerForm });
+  const { getByText, getByRole } = render(AppFixture, { socket, component: JoinGameForm });
 
   const button = getByText('Join game') as HTMLButtonElement;
   const dogAvatar = getByRole('radio', { name: '🐶' }) as HTMLInputElement;
